[nanvix-bench] E: Reduce feature gates
Refactor nanvix-bench to reduce compile-time feature gates from ~65
down to 7 by reorganizing the source tree and centralizing feature
validation.

Source tree:
- Split benchmarks/ into vmm/ (boot_time, snapshot_restore,
  warm_start_vmm) and system/ (cold_start, concurrent,
  echo_breakdown, round_trip_latency, warm_start) subdirectories.
- Move nanvixd helpers (setup, start, kill, cleanup,
  run_user_vm_echo_once) from benchmarks/mod.rs into
  benchmarks/system/mod.rs.
- Top-level benchmarks/mod.rs now only holds shared constants.

BenchmarkFlavour (benchmark.rs):
- Remove all #[cfg] gates from enum variants, Display, FromStr,
  get_program(), and is_l2().
- Add deployment(), requires_timestamp_messages(), and needs_nanvixd()
  methods to encode feature requirements as data rather than gates.

Dispatch (main.rs):
- Add validate_benchmark() using cfg!() macro to check feature
  compatibility at a single point before dispatch.
- Replace 13 feature-gated match arms and 12 nested
  timestamp-messages gate pairs with ungated match using | patterns
  and the deployment() method.

CLI (args.rs):
- Replace 4 mutually exclusive #[cfg] help-text blocks with dynamic
  generation using cfg!() runtime checks.
- Remove feature gates from concurrent and netns-pool-size validation.

Remaining gates (7 total):
- 3x gdb_port field in UserVmArgs (structural, requires gdb feature).
- 2x echo_breakdown impl + 1x stub + 1x module declaration (requires
  profiler crate from timestamp-messages feature).

No CLI or behavioral changes. All existing benchmark names and flags
are preserved.
