+++ src/utils/nanvix-bench/src/args.rs
@@ -40,49 +40,61 @@ impl Args {
     const OPT_TMP_DIR: &'static str = "-tmp-dir";
 
     fn usage(program_name: &str) {
-        #[cfg(all(not(feature = "l2"), feature = "multi-process"))]
-        let benchmarks = "\
-  boot-time              Measure raw user VM boot latency.
-  cold-start             Measure start-up latency from client's perspective.
-  cold-start-uvm         Measure start-up latency of the user VM only, excluding linuxd.
-  concurrent             Measure cold-start times as we increase the number of concurrent user VMs.
-  echo-breakdown         Analyze the latency contributions of each step in the data path.
-  round-trip-latency     Measure latency (warm-start) as we increase the payload size.
-  snapshot-restore       Measure snapshot restore latency vs boot-time.
-  warm-start             Measure round-trip latency from client's perspective.
-  warm-start-vmm         Measure raw round-trip latency inside the user VM.";
+        let mut benchmarks = String::new();
 
-        #[cfg(all(
-            not(feature = "l2"),
-            not(feature = "multi-process"),
-            feature = "single-process"
-        ))]
-        let benchmarks = "\
-  boot-time              Measure raw user VM boot latency.
+        // VMM-level benchmarks are always available.
+        benchmarks.push_str(
+            "\
+  boot-time              Measure raw user VM boot latency.\n",
+        );
+
+        // System-level benchmarks require multi-process or single-process.
+        if cfg!(any(feature = "multi-process", feature = "single-process")) {
+            benchmarks.push_str(
+                "\
   cold-start             Measure start-up latency from client's perspective.
   cold-start-uvm         Measure start-up latency of the user VM only, excluding linuxd.
   echo-breakdown         Analyze the latency contributions of each step in the data path.
-  round-trip-latency     Measure latency (warm-start) as we increase the payload size.
-  snapshot-restore       Measure snapshot restore latency vs boot-time.
-  warm-start             Measure round-trip latency from client's perspective.
-  warm-start-vmm         Measure raw round-trip latency inside the user VM.";
+  round-trip-latency     Measure latency (warm-start) as we increase the payload size.\n",
+            );
+        }
 
-        #[cfg(all(
-            not(feature = "l2"),
-            not(feature = "multi-process"),
-            not(feature = "single-process")
-        ))]
-        let benchmarks = "\
-  boot-time              Measure raw user VM boot latency.
-  snapshot-restore       Measure snapshot restore latency vs boot-time.
-  warm-start-vmm         Measure raw round-trip latency inside the user VM.";
+        // concurrent requires multi-process specifically.
+        if cfg!(feature = "multi-process") {
+            benchmarks.push_str(
+                "\
+  concurrent             Measure cold-start times as we increase the number of concurrent user \
+                 VMs.\n",
+            );
+        }
 
-        #[cfg(feature = "l2")]
-        let benchmarks = "\
+        // L2 variants.
+        if cfg!(feature = "l2") {
+            benchmarks.push_str(
+                "\
   cold-start-l2          Same as cold-start, but deploy linuxd inside an L2 VM.
   concurrent-l2          Same as concurrent, but deploy linuxd inside an L2 VM.
   echo-breakdown-l2      Same as echo-breakdown, but deploy linuxd inside L2 VM.
-  warm-start-l2          Same as warm-start, but deploy linuxd inside an L2 VM.";
+  warm-start-l2          Same as warm-start, but deploy linuxd inside an L2 VM.\n",
+            );
+        }
+
+        benchmarks.push_str(
+            "\
+  snapshot-restore       Measure snapshot restore latency vs boot-time.\n",
+        );
+
+        if cfg!(any(feature = "multi-process", feature = "single-process")) {
+            benchmarks.push_str(
+                "\
+  warm-start             Measure round-trip latency from client's perspective.\n",
+            );
+        }
+
+        benchmarks.push_str(
+            "\
+  warm-start-vmm         Measure raw round-trip latency inside the user VM.",
+        );
 
         println!(
             "\
@@ -229,28 +241,7 @@ Examples:
             Ok(benchmark) => {
                 match benchmark {
                     // The concurrent benchmarks take slightly different command-line arguments.
-                    #[cfg(all(feature = "multi-process", feature = "l2"))]
                     BenchmarkFlavour::Concurrent | BenchmarkFlavour::ConcurrentL2 => {
-                        // Must pass -num-concurrent-vms
-                        if num_concurrent_vms.is_none() {
-                            Self::usage(args[0].as_str());
-                            return Err(anyhow::anyhow!(
-                                "missing value for: {}",
-                                Self::OPT_NUM_CONCURRENT_VMS
-                            ));
-                        }
-
-                        // Must not pass -hwloc.
-                        if hwloc_file.is_some() {
-                            Self::usage(args[0].as_str());
-                            return Err(anyhow::anyhow!(
-                                "{benchmark} benchmark does not take {} flag",
-                                Self::OPT_HWLOC,
-                            ));
-                        }
-                    },
-                    #[cfg(all(feature = "multi-process", not(feature = "l2")))]
-                    BenchmarkFlavour::Concurrent => {
                         // Must pass -num-concurrent-vms
                         if num_concurrent_vms.is_none() {
                             Self::usage(args[0].as_str());
@@ -283,7 +274,6 @@ Examples:
                 // Reject -netns-pool-size when it would be silently ignored.
                 if netns_pool_size.is_some() {
                     match benchmark {
-                        #[cfg(feature = "l2")]
                         BenchmarkFlavour::ConcurrentL2 => {},
                         _ => {
                             Self::usage(args[0].as_str());
@@ -297,7 +287,6 @@ Examples:
 
                 // Derive netns pool size from the benchmark flavour.
                 let netns_pool_size = match benchmark {
-                    #[cfg(feature = "l2")]
                     BenchmarkFlavour::ConcurrentL2 => {
                         Some(netns_pool_size.unwrap_or(Self::DEFAULT_NETNS_POOL_SIZE))
                     },
