+++ src/utils/nanvix-bench/src/benchmarks/system/echo_breakdown.rs
@@ -1,7 +1,7 @@
 // Copyright(c) The Maintainers of Nanvix.
 // Licensed under the MIT License.
 
-use super::WARMUP_SLEEP_DURATION;
+use super::super::WARMUP_SLEEP_DURATION;
 use crate::benchmark::{
     Benchmark,
     LinuxdDeployment,
