+++ src/utils/nanvix-bench/src/benchmarks/vmm/mod.rs
@@ -1,0 +1,10 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Modules
+//==================================================================================================
+
+mod boot_time;
+mod snapshot_restore;
+mod warm_start_vmm;
