+++ src/utils/nanvix-bench/src/benchmarks/vmm/warm_start_vmm.rs
@@ -1,7 +1,7 @@
 // Copyright(c) The Maintainers of Nanvix.
 // Licensed under the MIT License.
 
-use super::{
+use super::super::{
     CHANNEL_CAPACITY,
     CLEANUP_SLEEP_DURATION,
     DEFAULT_PAYLOAD_SIZE,
